Fix macos image resizing crashes

+++ pad/generic/border.cpp
@@ -40,9 +40,8 @@ software in general.
 // slot (see resource.h / resource.cpp)
 //
 
-Pad_Border::~Pad_Border()
-{
-    Delete_per_display(); 
+Pad_Border::~Pad_Border() {
+    Delete_per_display();
 }
 
 //
@@ -51,15 +50,13 @@ Pad_Border::~Pad_Border()
 //
 
 void *
-Pad_Border::Alloc(Pad_Display *dpy)
-{
-    return (void*)dpy->Alloc_border(r, g, b);
+Pad_Border::Alloc(Pad_Display *dpy) {
+    return (void *) dpy->Alloc_border(r, g, b);
 }
 
 void
-Pad_Border::Free(Pad_Display *dpy, void *value)
-{
-    dpy->Free_border((Pad_3DBorder)value);
+Pad_Border::Free(Pad_Display *dpy, void *value) {
+    dpy->Free_border((Pad_3DBorder) value);
 }
 
 
@@ -78,83 +75,78 @@ static Pad_HashTable rgbtable;
 // the count reaches zero.
 //
 static void
-_Release(Pad_Border *border)
-{
+_Release(Pad_Border *border) {
     if (border && border->Deref()) {
-	if (border->Name()) {
-	    nametable.Remove((void*)border->Name());
-	} else {
-	    int red, green, blue;
-	    border->Get(red, green, blue);
-	    rgbtable.Remove((void*)CC_RGBA(red, green, blue, 0));
-	}
-	delete border;	    	    
-	border = NULL;
+        if (border->Name()) {
+            nametable.Remove((void *) border->Name());
+        } else {
+            intptr_t red, green, blue;
+            border->Get(red, green, blue);
+            rgbtable.Remove((void *) CC_RGBA(red, green, blue, 0));
+        }
+        delete border;
+        border = NULL;
     }
 }
 
 #define RELEASE() { _Release(border); border = NULL; }
 
-Pad_BorderRef::~Pad_BorderRef()
-{
+Pad_BorderRef::~Pad_BorderRef() {
     RELEASE();
 }
 
 void
-Pad_BorderRef::Set(int r, int g, int b)
-{
-    void *key = (void*)CC_RGBA(r, g, b, 0);
+Pad_BorderRef::Set(intptr_t r, intptr_t g, intptr_t b) {
+    void *key = (void *) CC_RGBA(r, g, b, 0);
 
     RELEASE();   // unreference old border
 
-    border = (Pad_Border *)rgbtable.Get(key);
+    border = (Pad_Border *) rgbtable.Get(key);
 
     if (!border) {
-	// need to allocate new border
-	border = new Pad_Border(r, g, b);
-	rgbtable.Set(key, border);
+        // need to allocate new border
+        border = new Pad_Border(r, g, b);
+        rgbtable.Set(key, border);
     }
 
     border->Addref();
 }
 
 void
-Pad_BorderRef::Set(const char *name)
-{
-    int red, green, blue;
+Pad_BorderRef::Set(const char *name) {
+    intptr_t red, green, blue;
 
-    RELEASE();		// unreference old border
+    RELEASE();        // unreference old border
 
     if (_Pad_colorname_to_rgb(name, red, green, blue)) {
-	if (name[0] == '#') {
-	    Set(red, green, blue);
-	} else {
-	    void *key = (void*)Pad_GetUid(name);
-	    border = (Pad_Border *)nametable.Get(key);
-	    
-	    if (!border) {
-		// need to allocate new color
-		border = new Pad_Border(name);
-		nametable.Set(key, border);
-	    }    	    
-	    border->Addref();
-	}
+        if (name[0] == '#') {
+            Set(red, green, blue);
+        } else {
+            void *key = (void *) Pad_GetUid(name);
+            border = (Pad_Border *) nametable.Get(key);
+
+            if (!border) {
+                // need to allocate new color
+                border = new Pad_Border(name);
+                nametable.Set(key, border);
+            }
+            border->Addref();
+        }
     }
 }
 
 void
-Pad_BorderRef::Set(const Pad_Color *otherColor)
-{
+Pad_BorderRef::Set(const Pad_Color *otherColor) {
     RELEASE();
 
     if (otherColor) {
-	if (otherColor->Name()) {
-	    Set(otherColor->Name());
-	} else {
-	    int red, green, blue;
-	    otherColor->Get(red, green, blue);
-	    Set(red, green, blue);
-	}
+        if (otherColor->Name()) {
+            Set(otherColor->Name());
+        } else {
+            intptr_t red, green, blue;
+            otherColor->Get(red, green, blue);
+            Set(red, green, blue);
+        }
     }
 }
 
@@ -162,33 +154,30 @@ Pad_BorderRef::Set(const Pad_Color *otherColor)
 // Sets border to either black or white...
 //
 void
-Pad_BorderRef::Set_contrasting(int red, int green, int blue)
-{
+Pad_BorderRef::Set_contrasting(intptr_t red, intptr_t green, intptr_t blue) {
     if ((.30 * red + .61 * green + .11 * blue) <= 127) {
-	// other color is dark - so use white
-	Set(255, 255, 255);
+        // other color is dark - so use white
+        Set(255, 255, 255);
     } else {
-    	// use black
-    	Set(0, 0, 0);
+        // use black
+        Set(0, 0, 0);
     }
 }
 
 void
-Pad_BorderRef::Get(int &red, int &green, int &blue) const
-{
+Pad_BorderRef::Get(intptr_t &red, intptr_t &green, intptr_t &blue) const {
     if (border)
-      border->Get(red, green, blue);
+        border->Get(red, green, blue);
     else
-      red = green = blue = 0;
+        red = green = blue = 0;
 }
 
 void
-Pad_BorderRef::Get(Pad_String &name) const
-{
+Pad_BorderRef::Get(Pad_String &name) const {
     if (border) {
-	border->Get(name);
+        border->Get(name);
     } else {
-	name = "none";
+        name = "none";
     }
 }
 
@@ -199,43 +188,40 @@ Pad_BorderRef::Get(Pad_String &name) const
 //
 
 Pad_Border *
-Pad_LookupBorderByName(char *name)
-{
+Pad_LookupBorderByName(char *name) {
     if (name && name[0] == '#') {
-	// uses #rrggbb encoding - look up color in rgb table.
-	int red, green, blue;
-	_Pad_colorname_to_rgb(name, red, green, blue);
-	return (Pad_Border*)rgbtable.Get((void*)CC_RGBA(red, green, blue, 0));
+        // uses #rrggbb encoding - look up color in rgb table.
+        intptr_t red, green, blue;
+        _Pad_colorname_to_rgb(name, red, green, blue);
+        return (Pad_Border *) rgbtable.Get((void *) CC_RGBA(red, green, blue, 0));
     } else {
-	// an X color name - look it up in the name table.
-	return (Pad_Border*)nametable.Get((void*)Pad_GetUid(name));
+        // an X color name - look it up in the name table.
+        return (Pad_Border *) nametable.Get((void *) Pad_GetUid(name));
     }
 }
 
 Pad_Border *
-Pad_AllocBorderByName(char *name)
-{
+Pad_AllocBorderByName(char *name) {
     Pad_Border *border = Pad_LookupBorderByName(name);
 
     if (!border) {
-	border = new Pad_Border(name);
-	if (border->Name() == NULL) {
-	    // use rgb table
-	    int red, green, blue;
-	    border->Get(red, green, blue);
-	    rgbtable.Set((void*)CC_RGBA(red, green, blue, 0), (void*)border);
-	} else {
-	    // use name table
-	    nametable.Set((void*)border->Name(), (void*)border);
-	}
+        border = new Pad_Border(name);
+        if (border->Name() == NULL) {
+            // use rgb table
+            intptr_t red, green, blue;
+            border->Get(red, green, blue);
+            rgbtable.Set((void *) CC_RGBA(red, green, blue, 0), (void *) border);
+        } else {
+            // use name table
+            nametable.Set((void *) border->Name(), (void *) border);
+        }
     }
     border->Addref();
     return border;
 }
 
 int
-Pad_FreeBorderByName(char *name)
-{
+Pad_FreeBorderByName(char *name) {
     Pad_Border *border = Pad_LookupBorderByName(name);
     _Release(border);
     return (border ? TRUE : FALSE);
@@ -247,34 +233,33 @@ Pad_FreeBorderByName(char *name)
 // Return TRUE if valid relief type, or FALSE otherwise.
 //
 Pad_Bool
-Pad_Relief_type_to_name(int relief, Pad_String &borderReliefName)
-{
+Pad_Relief_type_to_name(int relief, Pad_String &borderReliefName) {
     Pad_Bool rc = TRUE;
 
     switch (relief) {
-      case PAD_RELIEF_FLAT:
-	borderReliefName = "flat";
-	break;
-      case PAD_RELIEF_RAISED:
-	borderReliefName = "raised";
-	break;
-      case PAD_RELIEF_SUNKEN:
-	borderReliefName = "sunken";
-	break;
-      case PAD_RELIEF_RIDGE:
-	borderReliefName = "ridge";
-	break;
-      case PAD_RELIEF_GROOVE:
-	borderReliefName = "groove";
-	break;
-      default:
-        rc = FALSE;
-	Pad_errorString = "Invalid borderrelief type: ";
-	Pad_errorString += relief;
-	break;
+        case PAD_RELIEF_FLAT:
+            borderReliefName = "flat";
+            break;
+        case PAD_RELIEF_RAISED:
+            borderReliefName = "raised";
+            break;
+        case PAD_RELIEF_SUNKEN:
+            borderReliefName = "sunken";
+            break;
+        case PAD_RELIEF_RIDGE:
+            borderReliefName = "ridge";
+            break;
+        case PAD_RELIEF_GROOVE:
+            borderReliefName = "groove";
+            break;
+        default:
+            rc = FALSE;
+            Pad_errorString = "Invalid borderrelief type: ";
+            Pad_errorString += relief;
+            break;
     }
 
-    return(rc);
+    return (rc);
 }
 
 //
@@ -282,22 +267,21 @@ Pad_Relief_type_to_name(int relief, Pad_String &borderReliefName)
 // Return TRUE if valid relief type, or FALSE otherwise.
 //
 Pad_Bool
-Pad_Relief_name_to_type(char *reliefName, int &borderReliefType)
-{
+Pad_Relief_name_to_type(char *reliefName, int &borderReliefType) {
     if (!strcmp(reliefName, "flat")) {
-	borderReliefType = PAD_RELIEF_FLAT;
+        borderReliefType = PAD_RELIEF_FLAT;
     } else if (!strcmp(reliefName, "raised")) {
-	borderReliefType = PAD_RELIEF_RAISED;
+        borderReliefType = PAD_RELIEF_RAISED;
     } else if (!strcmp(reliefName, "sunken")) {
-	borderReliefType = PAD_RELIEF_SUNKEN;
+        borderReliefType = PAD_RELIEF_SUNKEN;
     } else if (!strcmp(reliefName, "groove")) {
-	borderReliefType = PAD_RELIEF_GROOVE;
+        borderReliefType = PAD_RELIEF_GROOVE;
     } else if (!strcmp(reliefName, "ridge")) {
-	borderReliefType = PAD_RELIEF_RIDGE;
+        borderReliefType = PAD_RELIEF_RIDGE;
     } else {
-	Pad_errorString = "Invalid borderrelief name: ";
-	Pad_errorString += reliefName;
-	return(FALSE);
+        Pad_errorString = "Invalid borderrelief name: ";
+        Pad_errorString += reliefName;
+        return (FALSE);
     }
-    return(TRUE);
+    return (TRUE);
 }
